fix: Styled 파일과 types 파일 잘못 기재한 부분 수정했습니다.
- 파일명을 제대로 안 보고 작성해버렸습니다 죄송합니다.
