+++ src/components/Pagination/Pagination.tsx
@@ -1,10 +1,10 @@
 import { useMemo } from 'react';
 import { IoIosArrowBack, IoIosArrowForward } from 'react-icons/io';
-import { PaginationProps } from './Pagination.styled';
-import { StyledPageButton, StyledPaginationControl } from './Pagination.types';
+import { PaginationProps } from './Pagination.types';
+import { StyledPageButton, StyledPaginationControl } from './Pagination.styled';
 
 export const Pagination = ({ limit, currentPage, onClick: handleClick, totalResults }: PaginationProps) => {
-  
+
   const { pageStartNum, pageEndNum } = useMemo(() => {
     const lastPageNum = Math.ceil(totalResults / limit);
     const pageStartNum = Math.max(currentPage - 2, 1);
